+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GP2-2856  remove unused api calls
 
 ### Fixed bugs
+- GP2-2887 - DAC_Focus_Order_04 - EP data snapshot fix
 - GP2-2886 - DAC_Focus_Order_03 - EP Objectives list focus management
 - GP2-2885 - DAC_Focus_Order_02 - Dynamic aria label and focus on WTM page
 - GP2-2884 - DAC Focus Order 01 - tabindex on mobile header
@@ -23,7 +24,9 @@
 - GP2-2916 - DAC_CSS_Content_01
 - GP2-2917 - DAC_CSS_Content_02
 - GP2-2918 - DAC_CSS_Content_03
-- GP2-2919 -DAC_CSS_Content_04
+- GP2-2919 - DAC_CSS_Content_04
+- GP2-2897 - DAC_Name_Role_Value_01
+- GP2-2899 - DAC_Name_Role_Value_03
 
 ## [1.10.1](https://github.com/uktrade/great-cms/releases/tag/1.10.1)
 ### Enhancements